+++ tests/countingcuckoo_test.py
@@ -10,8 +10,8 @@ from probables import (CountingCuckooFilter, CuckooFilterFullError,
 class TestCountingCuckooFilter(unittest.TestCase):
     ''' base Cuckoo Filter test '''
 
-    def test_cuckoo_filter_default(self):
-        ''' test cuckoo filter default properties '''
+    def test_c_cuckoo_filter_default(self):
+        ''' test counting cuckoo filter default properties '''
         cko = CountingCuckooFilter()
         self.assertEqual(10000, cko.capacity)
         self.assertEqual(4, cko.bucket_size)
@@ -19,8 +19,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertEqual(2, cko.expansion_rate)
         self.assertEqual(True, cko.auto_expand)
 
-    def test_cuckoo_filter_diff(self):
-        ''' test cuckoo filter non-standard properties '''
+    def test_c_cuckoo_filter_diff(self):
+        ''' test counting cuckoo filter non-standard properties '''
         cko = CountingCuckooFilter(capacity=100, bucket_size=2, max_swaps=5,
                                    expansion_rate=4, auto_expand=False)
         self.assertEqual(100, cko.capacity)
@@ -29,8 +29,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertEqual(4, cko.expansion_rate)
         self.assertEqual(False, cko.auto_expand)
 
-    def test_cuckoo_filter_add(self):
-        ''' test adding to the cuckoo filter '''
+    def test_c_cuckoo_filter_add(self):
+        ''' test adding to the counting cuckoo filter '''
         cko = CountingCuckooFilter()
         cko.add('this is a test')
         self.assertEqual(cko.elements_added, 1)
@@ -39,8 +39,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         cko.add('this is yet another test')
         self.assertEqual(cko.elements_added, 3)
 
-    def test_cuckoo_filter_remove(self):
-        ''' test removing from the cuckoo filter '''
+    def test_c_cuckoo_filter_remove(self):
+        ''' test removing from the counting cuckoo filter '''
         cko = CountingCuckooFilter()
         cko.add('this is a test')
         self.assertEqual(cko.elements_added, 1)
@@ -48,16 +48,24 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertEqual(cko.elements_added, 2)
         cko.add('this is yet another test')
         self.assertEqual(cko.elements_added, 3)
+        self.assertEqual(cko.unique_elements, 3)
+        cko.add('this is a test')
+        cko.add('this is a test')
+        cko.add('this is a test')
+        self.assertEqual(cko.elements_added, 6)
+        self.assertEqual(cko.unique_elements, 3)
 
-        res = cko.remove('this is a test')
+        res = cko.remove('this is another test')
         self.assertTrue(res)
-        self.assertEqual(cko.elements_added, 2)
-        self.assertFalse(cko.check('this is a test'))
-        self.assertTrue(cko.check('this is another test'))
+        self.assertEqual(cko.elements_added, 5)
+        self.assertEqual(cko.unique_elements, 2)
+
+        self.assertTrue(cko.check('this is a test'))
+        self.assertFalse(cko.check('this is another test'))
         self.assertTrue(cko.check('this is yet another test'))
 
-    def test_cuckoo_filter_remove_miss(self):
-        ''' test removing from the cuckoo filter when not present '''
+    def test_c_cuckoo_filter_rmv_miss(self):
+        ''' test removing from the counting cuckoo filter when not present '''
         cko = CountingCuckooFilter()
         cko.add('this is a test')
         self.assertEqual(cko.elements_added, 1)
@@ -73,15 +81,15 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertTrue(cko.check('this is another test'))
         self.assertTrue(cko.check('this is yet another test'))
 
-    def test_cuckoo_filter_lots(self):
-        ''' test inserting lots into the cuckoo filter '''
+    def test_c_cuckoo_filter_lots(self):
+        ''' test inserting lots into the counting cuckoo filter '''
         cko = CountingCuckooFilter(capacity=100, bucket_size=2, max_swaps=100)
         for i in range(125):
             cko.add(str(i))
         self.assertEqual(cko.elements_added, 125)
 
-    def test_cuckoo_filter_full(self):
-        ''' test inserting until cuckoo filter is full '''
+    def test_c_cuckoo_filter_full(self):
+        ''' test inserting until counting cuckoo filter is full '''
         def runner():
             ''' runner '''
             cko = CountingCuckooFilter(capacity=100, bucket_size=2,
@@ -90,8 +98,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
                 cko.add(str(i))
         self.assertRaises(CuckooFilterFullError, runner)
 
-    def test_cuckoo_full_msg(self):
-        ''' test exception message for full cuckoo filter '''
+    def test_c_cuckoo_full_msg(self):
+        ''' test exception message for full counting cuckoo filter '''
         try:
             cko = CountingCuckooFilter(capacity=100, bucket_size=2,
                                        max_swaps=100, auto_expand=False)
@@ -103,8 +111,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         else:
             self.assertEqual(True, False)
 
-    def test_cuckoo_idx(self):
-        ''' test that the indexing works correctly for cuckoo filter swap '''
+    def test_c_cuckoo_idx(self):
+        ''' test that the indexing works correctly for counting cuckoo filter swap '''
         cko = CountingCuckooFilter(capacity=100, bucket_size=2, max_swaps=5)
         txt = 'this is a test'
         idx_1, idx_2, fingerprint = cko._generate_fingerprint_info(txt)
@@ -112,8 +120,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertEqual(idx_1, index_1)
         self.assertEqual(idx_2, index_2)
 
-    def test_cuckoo_filter_check(self):
-        ''' test checking if element in cuckoo filter '''
+    def test_c_cuckoo_filter_check(self):
+        ''' test checking if element in counting cuckoo filter '''
         cko = CountingCuckooFilter()
         cko.add('this is a test')
         cko.add('this is another test')
@@ -124,8 +132,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertEqual(cko.check('this is not another test'), False)
         self.assertEqual(cko.check('this is not a test'), False)
 
-    def test_cuckoo_filter_in(self):
-        ''' test checking using 'in' cuckoo filter '''
+    def test_c_cuckoo_filter_in(self):
+        ''' test checking using 'in' counting cuckoo filter '''
         cko = CountingCuckooFilter()
         cko.add('this is a test')
         cko.add('this is another test')
@@ -136,8 +144,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         self.assertEqual('this is not another test' in cko, False)
         self.assertEqual('this is not a test' in cko, False)
 
-    def test_cuckoo_filter_dup_add(self):
-        ''' test adding same item multiple times cuckoo filter '''
+    def test_c_cuckoo_filter_dup_add(self):
+        ''' test adding same item multiple times counting cuckoo filter '''
         cko = CountingCuckooFilter()
         cko.add('this is a test')
         cko.add('this is another test')
@@ -147,9 +155,10 @@ class TestCountingCuckooFilter(unittest.TestCase):
         cko.add('this is another test')
         cko.add('this is yet another test')
         self.assertEqual(cko.elements_added, 6)
+        self.assertEqual(cko.unique_elements, 3)
 
-    def test_cuckoo_filter_l_fact(self):
-        ''' test the load factor of the cuckoo filter '''
+    def test_c_cuckoo_filter_l_fact(self):
+        ''' test the load factor of the counting cuckoo filter '''
         cko = CountingCuckooFilter(capacity=100, bucket_size=2, max_swaps=10)
         self.assertEqual(cko.load_factor(), 0.0)
         for i in range(50):
@@ -159,8 +168,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
             cko.add(str(i + 50))
         self.assertEqual(cko.load_factor(), 0.50)
 
-    def test_cuckoo_filter_export(self):
-        ''' test exporting a cuckoo filter '''
+    def test_c_cuckoo_filter_export(self):
+        ''' test exporting a counting cuckoo filter '''
         def runner():
             ''' runner '''
             cko = CountingCuckooFilter()
@@ -178,8 +187,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         # self.assertEqual(md5sum, md5_out)
         # os.remove(filename)
 
-    def test_cuckoo_filter_load(self):
-        ''' test loading a saved cuckoo filter '''
+    def test_c_cuckoo_filter_load(self):
+        ''' test loading a saved counting cuckoo filter '''
         def runner():
             ''' runner '''
             CountingCuckooFilter(filepath='./test.cck')
@@ -203,8 +212,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
         # self.assertEqual(0.025, ckf.load_factor())
         # os.remove(filename)
 
-    def test_cuckoo_filter_expand_els(self):
-        ''' test out the expansion of the cuckoo filter '''
+    def test_c_cuckoo_filter_expand_els(self):
+        ''' test out the expansion of the counting cuckoo filter '''
         cko = CountingCuckooFilter()
         for i in range(200):
             cko.add(str(i))
@@ -213,8 +222,8 @@ class TestCountingCuckooFilter(unittest.TestCase):
             self.assertGreater(cko.check(str(i)), 0)
         self.assertEqual(20000, cko.capacity)
 
-    def test_cuckoo_filter_auto_expand(self):
-        ''' test inserting until cuckoo filter is full '''
+    def test_c_cuckoo_filter_auto_exp(self):
+        ''' test inserting until counting cuckoo filter is full '''
         cko = CountingCuckooFilter(capacity=100, bucket_size=2, max_swaps=100)
         for i in range(375):  # this would fail if it doesn't expand
             cko.add(str(i))
@@ -223,9 +232,25 @@ class TestCountingCuckooFilter(unittest.TestCase):
         for i in range(375):
             self.assertGreater(cko.check(str(i)), 0)
 
-    def test_cuckoo_filter_bin(self):
+    def test_c_cuckoo_filter_bin(self):
         ''' test the cuckoo bin repr '''
         cko = CountingCuckooFilter(capacity=1, bucket_size=2, max_swaps=100)
         cko.add('this is a test')
         self.assertEqual('[(fingerprint:3057276164 count:1)]',
                          str(cko.buckets[0]))
+
+
+    def test_c_cuckoo_filter_str(self):
+        ''' test the str representation of the counting cuckoo filter '''
+        cko = CountingCuckooFilter(capacity=100, bucket_size=2, max_swaps=100)
+        for i in range(75):
+            cko.add(str(i))
+        msg = ('CountingCuckooFilter:\n'
+               '\tCapacity: 100\n'
+               '\tTotal Bins: 200\n'
+               '\tLoad Factor: 37.5%\n'
+               '\tInserted Elements: 75\n'
+               '\tMax Swaps: 100\n'
+               '\tExpansion Rate: 2\n'
+               '\tAuto Expand: True')
+        self.assertEqual(str(cko), msg)